Feat: 날씨 응답 모델 대체
- WeatherModel 제거 -> WeatherResponse 추가
